@@ -3,6 +3,7 @@
   "version": "1.0.3",
   "description": "@ModernPoacher/modernpoacher",
   "private": true,
+  "type": "module",
   "author": "Jonathan Perry for Modern Poacher Limited <modernpoacher@modernpoacher.com>",
   "scripts": {
     "build:clean": "cross-env NODE_ENV=production concurrently -r \"npm run build:js:clean\" \"npm run build:css:clean\"",
@@ -16,7 +17,7 @@
     "build:css:watch": "gulp build:css:watch",
     "lint": "npm run lint:js && npm run lint:css",
     "lint:fix": "npm run lint:js:fix && npm run lint:css:fix",
-    "lint:js": "cross-env NODE_ENV=production eslint .",
+    "lint:js": "cross-env NODE_ENV=production eslint . --ext .mjs --ext .cjs --ext .js",
     "lint:js:fix": "npm run lint:js -- --fix",
     "lint:css": "cross-env NODE_ENV=production stylelint src/sass",
     "lint:css:fix": "npm run lint:css -- --fix",
@@ -27,21 +28,21 @@
     "debug": "^4.3.4"
   },
   "devDependencies": {
-    "@babel/cli": "^7.18.6",
-    "@babel/core": "^7.18.6",
-    "@babel/eslint-parser": "^7.18.2",
-    "@babel/preset-env": "^7.18.6",
-    "@babel/register": "^7.18.6",
+    "@babel/cli": "^7.18.9",
+    "@babel/core": "^7.18.9",
+    "@babel/eslint-parser": "^7.18.9",
+    "@babel/preset-env": "^7.18.9",
+    "@babel/register": "^7.18.9",
     "autoprefixer": "^10.4.7",
     "babel-loader": "^8.2.5",
     "babel-plugin-minify-dead-code-elimination": "^0.5.2",
     "babel-plugin-module-resolver": "^4.1.0",
     "clean-webpack-plugin": "^4.0.0",
-    "concurrently": "^7.2.2",
-    "core-js": "^3.23.5",
+    "concurrently": "^7.3.0",
+    "core-js": "^3.24.0",
     "cross-env": "^7.0.3",
     "cssnano": "^5.1.12",
-    "del": "^6.1.1",
+    "del": "^7.0.0",
     "eslint": "^8.20.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-import-resolver-babel-module": "^5.3.1",
@@ -60,13 +61,18 @@
     "postcss": "^8.4.14",
     "postcss-scss": "^4.0.4",
     "rimraf": "^3.0.2",
-    "sass": "^1.53.0",
+    "sass": "^1.54.0",
     "stylelint": "^14.9.1",
     "stylelint-config-recommended-scss": "^7.0.0",
     "stylelint-scss": "^4.3.0",
     "terser-webpack-plugin": "^5.3.3",
     "vinyl-paths": "3.0.1",
-    "webpack": "^5.73.0",
+    "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0"
+  },
+  "imports": {
+    "#build/gulp": "./build/gulp/index.mjs",
+    "#build/gulp/css-from-sass": "./build/gulp/css-from-sass.mjs",
+    "#build/gulp/handle-watch-error": "./build/gulp/handle-watch-error.mjs"
   }
 }
